Add loading state for ListPrompts and implement cache for RouterView
This update adds a loading state to ListPrompts.vue, which shows an animated spinner during the loading of list documents. This enhancement helps in visually indicating the user about a pending operation and improves user experience by making our application feel more responsive and interactive.

In addition to this, 'keep-alive' tag is included in App.vue to cache the RouterView components. This preserves the state of the current component when navigating throughout the application thus improving performance and usability.